test: Move label layout fixtures out of source
